Refactor: Consolidate context types in types.ts
- Add BaseTransformContext interface with shared signals/memos properties
- Move JsxToIRContext from jsx-to-ir.ts to types.ts
- Move ServerJsxContext from ir-to-server-jsx.ts to types.ts
- Move CollectContext from ir-to-client-js.ts to types.ts
- Move ClientJsGeneratorContext from jsx-compiler.ts to types.ts
- Move CompilerWarning type to types.ts
- Update imports and add re-exports for backwards compatibility
